[update] 현재 상품의 이미지가 비어있다면 기존 이미지 삭제 로직을 실행하지 않도록 변경

+++ src/main/java/com/example/baglemonster/product/service/ProductService.java
@@ -81,7 +81,8 @@ public class ProductService {
         String currentPictureUrl = product.getProductPictureUrl();
         String productPictureUrl = currentPictureUrl;
         if (file != null) {
-            s3UploadService.deleteFile(currentPictureUrl);
+            if(currentPictureUrl != null)
+                s3UploadService.deleteFile(currentPictureUrl);
             productPictureUrl  = s3UploadService.uploadFile(file);
         }
